% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{add_numbers_at_length}
\alias{add_numbers_at_length}
\title{Add numbers-at-length to a DATRAS object}
\usage{
add_numbers_at_length(x, cm_breaks = NULL, by = 1)
}
\arguments{
\item{x}{A \code{DATRASraw} object.}

\item{cm_breaks}{Numeric vector of break points defining the length classes
in cm. If \code{NULL}, break points are created automatically from the observed
range of \code{HL$LngtCm} using \code{by}.}

\item{by}{Numeric scalar giving the width of the default length classes in
cm. Default is 1cm (for herring, sprat and other smaller species 0.5 might
be more appropriate). Only used when \code{cm_breaks = NULL}.}
}
\value{
A \code{DATRASraw} object with numbers-at-length added to the \code{HH}
component as matrix \code{N}.
}
\description{
Calculates numbers-at-length for each haul and stores the result in the
\code{HH} table of a \code{DATRASraw} object as matrix \code{N}, with one row per haul and
one column per length class.
}
\details{
By default, numbers-at-length are calculated using the available haul-level
length data in \code{HL} and the default length recording resolution of the data.

The function first adds numbers-at-length at the finest available length
resolution and stores the result as \code{x[["HH"]][["N"]]}.

The applied break points are stored as attribute \code{cm.breaks} on the returned
object.
}
\examples{
## Add numbers-at-length using the default length resolution
x <- add_numbers_at_length(dab)

}
\seealso{
\code{\link[=add_total_numbers_by_haul]{add_total_numbers_by_haul()}}
}
