% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{add_weight_by_haul}
\alias{add_weight_by_haul}
\title{Add total haul biomass to \code{HH}}
\usage{
add_weight_by_haul(x, per_minute = TRUE)
}
\arguments{
\item{x}{A \code{datras_raw} object.}

\item{per_minute}{Logical. If \code{TRUE} (default), biomass is divided by haul
duration in minutes.}
}
\value{
A \code{datras_raw} object with haul-level biomass added to \code{HH}.
}
\description{
Calculate total biomass by haul and add it to the \code{HH} component of a
\code{datras_raw} / \code{DATRASraw} object using the underlying DATRAS method.
}
\details{
This function is a thin wrapper around \code{\link[DATRAS:addWeightByHaul]{DATRAS::addWeightByHaul()}}, applied
after checking that the input is a valid \code{datras_raw} object.
}
\examples{
\dontrun{
x <- add_weight_by_haul(x)
x <- add_weight_by_haul(x, per_minute = FALSE)
}

}
\seealso{
\code{\link[=add_weight]{add_weight()}}, \code{\link[=add_weight_by_haul_empirical]{add_weight_by_haul_empirical()}}
}
