% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{add_weight_by_haul_empirical}
\alias{add_weight_by_haul_empirical}
\title{Add empirical total haul biomass to \code{HH}}
\usage{
add_weight_by_haul_empirical(x, per_minute = TRUE)
}
\arguments{
\item{x}{A \code{datras_raw} object.}

\item{per_minute}{Logical. If \code{TRUE} (default), haul biomass is divided by
haul duration in minutes.}
}
\value{
A \code{datras_raw} object with an added \code{HaulWgt} field in \code{HH}.
}
\description{
Calculate total biomass by haul from empirical species-specific
length-weight parameters and add it to the \code{HH} component of a
\code{datras_raw} / \code{DATRASraw} object.
}
\details{
The function uses the Aphia ID in the \code{HL} table to look up the empirical
length-weight parameters \code{a} and \code{b} in \code{species_info}, predicts weight for
each length class defined by \code{attr(x, "cm.breaks")}, and combines these
with haul-specific numbers-at-length stored in \code{HH$N} to compute total haul
biomass.

The function requires that exactly one unique \code{Valid_Aphia} value is present
in \code{x[["HL"]]}, and that matching empirical length-weight parameters are
available in the internal \code{species_info} table.

Weight-at-length is calculated as:
\deqn{
  W = a \times L^b
}

Total biomass for each haul is then calculated by multiplying the haul-level
numbers-at-length by the predicted weight-at-length vector and summing across
length classes.
}
\examples{
\dontrun{
x <- add_weight_by_haul_empirical(x)
x <- add_weight_by_haul_empirical(x, per_minute = FALSE)
}

}
\seealso{
\code{\link[=add_weight]{add_weight()}}, \code{\link[=add_weight_empirical]{add_weight_empirical()}}, \code{\link[=add_weight_by_haul]{add_weight_by_haul()}}
}
