% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{add_weight_empirical}
\alias{add_weight_empirical}
\title{Add empirical weight-at-length estimates to \code{HH}}
\usage{
add_weight_empirical(x, per_minute = TRUE)
}
\arguments{
\item{x}{A \code{datras_raw} object.}

\item{per_minute}{Logical. If \code{TRUE} (default), estimated weights are divided
by haul duration in minutes.}
}
\value{
A \code{datras_raw} object with an added \code{Wgt} field in \code{HH}.
}
\description{
Calculate weight-at-length from empirical species-specific length-weight
parameters and add the resulting weights to the \code{HH} component of a
\code{datras_raw} / \code{DATRASraw} object.
}
\details{
The function uses the Aphia ID in the \code{HL} table to look up the empirical
length-weight parameters \code{a} and \code{b} in \code{species_info}, predicts weight for
each length class defined by \code{attr(x, "cm.breaks")}, and multiplies these
by numbers-at-length stored in \code{HH$N}.

The function requires that exactly one unique \code{Valid_Aphia} value is present
in \code{x[["HL"]]}, and that matching empirical length-weight parameters are
available in the internal \code{species_info} table.

Weight-at-length is calculated as:
\deqn{
  W = a \times L^b
}

where \code{L} is length in centimetres and \code{W} is weight in grams.

The resulting length-specific weights are multiplied by the haul-specific
numbers-at-length matrix \code{HH$N}. If \code{per_minute = TRUE}, the resulting
weights are standardized by haul duration.
}
\examples{
\dontrun{
x <- add_weight_empirical(x)
x <- add_weight_empirical(x, per_minute = FALSE)
}

}
\seealso{
\code{\link[=add_weight]{add_weight()}}, \code{\link[=add_weight_by_haul_empirical]{add_weight_by_haul_empirical()}}
}
