% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swept_area.R
\name{addSweptAreaSimple}
\alias{addSweptAreaSimple}
\title{Add Swept Area based on wing spread / beam width using a simple approach.}
\usage{
addSweptAreaSimple(
  d,
  minSpeed = 1,
  minDist = 0,
  maxDistDev = 0.2,
  impute.missing = FALSE
)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
DATRASraw object with two SweptArea indices: SweptArea and SweptArea.median. The latter assumes a fixed trawl width for each gear category (the median).
}
\description{
Add Swept Area based on wing spread / beam width using a simple approach.
}
