% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{addWeight}
\alias{addWeight}
\title{Calculate weight by length classes and add to HH-records}
\usage{
addWeight(d, to1min = TRUE, maxL = NULL, maxW = NULL)
}
\arguments{
\item{d}{DATRASraw object}

\item{to1min}{divide by haul duration in minutes? (defaults to TRUE)}

\item{maxL}{x}

\item{maxW}{x}
}
\value{
DATRASraw object
}
\description{
Calculate weight by length classes and add to HH-records
}
