% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{addWeightByHaulEmpirical}
\alias{addWeightByHaulEmpirical}
\title{Calculate total biomass by haul using a and b, and add to HH-records}
\usage{
addWeightByHaulEmpirical(d, to1min = TRUE)
}
\arguments{
\item{d}{DATRASraw object}

\item{to1min}{divide by haul duration in minutes? (defaults to TRUE)}
}
\value{
DATRASraw object
}
\description{
Calculate total biomass by haul using empirical a and b, and add to HH-records
}
