% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishglob.R
\name{addWeightFishglob}
\alias{addWeightFishglob}
\title{calculate weight by length classes and add to hh records}
\usage{
addWeightFishglob(x)
}
\arguments{
\item{x}{a DATRASraw object}
}
\value{
a DATRASraw object with hl table updated with weight fields
}
\description{
calculate weight by length classes using empirical a and b, and add to
hh records according to fishglob workflow
}
\details{
\itemize{
\item weight-at-length is calculated per individual using empirical a and b parameters
\item total weight per length class is computed as number at length * individual weight
\item optionally, weight can be divided by haul duration in minutes
\item final weight is returned in kg
}
}
