% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{checkLength}
\alias{checkLength}
\title{Create and inspect length spectrum}
\usage{
checkLength(
  x,
  cm.breaks = seq(min(x[[3]]$LngtCm, na.rm = TRUE), max(x[[3]]$LngtCm, na.rm = TRUE) +
    by, by = by),
  by = NULL,
  length.percentile = 99.9999,
  plot = TRUE
)
}
\arguments{
\item{x}{DATRASraw object}

\item{cm.breaks}{x}

\item{by}{If NULL, most coarse accuracy is used (using DATRAS internal
function getAccuracyCM)}

\item{length.percentile}{x}

\item{plot}{TRUE}
}
\value{
N at length matrix
}
\description{
Create and inspect length measurements (in HL)
}
