% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{checkWeight}
\alias{checkWeight}
\title{Check weight information (in CA)}
\usage{
checkWeight(d, maxL = NULL, maxW = NULL, plot = TRUE)
}
\arguments{
\item{d}{DATRASraw object}

\item{maxL}{x}

\item{maxW}{x}

\item{plot}{TRUE}
}
\value{
Info
}
\description{
Check weight information (in CA)
}
