% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean data}
\usage{
clean(
  x,
  aphias = NULL,
  years = NULL,
  quarters = NULL,
  gears = NULL,
  impute.missing.depth = TRUE,
  correct.species = TRUE
)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
Cleaned DATRASraw object.
}
\description{
Clean data
}
