% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{correctSpecies}
\alias{correctSpecies}
\title{Correct species names}
\usage{
correctSpecies(x)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
An object of DATRASraw object
}
\description{
Correct species names
}
