% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadDATRAS}
\alias{downloadDATRAS}
\title{Download DATRAS survey information}
\usage{
downloadDATRAS(
  surveys = NULL,
  years = NULL,
  download.missing.only = TRUE,
  dir = NULL
)
}
\arguments{
\item{surveys}{surveys}
}
\description{
Download DATRAS survey information
}
