% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_prune.R
\name{drop_ca}
\alias{drop_ca}
\title{Remove the CA table from a \code{datras_raw} object}
\usage{
drop_ca(x)
}
\arguments{
\item{x}{A \code{datras_raw} object.}
}
\value{
The input object with \code{CA} set to \code{NULL}.
}
\description{
Set the \code{CA} (biological sampling) table of a \code{datras_raw} object to \code{NULL}.
}
\examples{
\dontrun{
x_no_ca <- drop_ca(x)
}
}
\seealso{
\code{\link[=drop_hl]{drop_hl()}}, \code{\link[=prune_datras]{prune_datras()}}
}
