% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_prune.R
\name{drop_hl}
\alias{drop_hl}
\title{Remove the HL table from a \code{datras_raw} object}
\usage{
drop_hl(x)
}
\arguments{
\item{x}{A \code{datras_raw} object.}
}
\value{
The input object with \code{HL} set to \code{NULL}.
}
\description{
Set the \code{HL} (length-frequency) table of a \code{datras_raw} object to \code{NULL}.
}
\examples{
\dontrun{
x_no_hl <- drop_hl(x)
}
}
\seealso{
\code{\link[=drop_ca]{drop_ca()}}, \code{\link[=prune_datras]{prune_datras()}}
}
