% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format.datras_raw}
\alias{format.datras_raw}
\title{Format DATRAS data}
\usage{
\method{format}{datras_raw}(x, ...)
}
\arguments{
\item{x}{a datras_raw object.}

\item{...}{extra options}
}
\value{
Dataframe in format x.
}
\description{
Format DATRAS data
}
