% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format.DATRASraw}
\alias{format.DATRASraw}
\title{Format DATRAS data}
\usage{
\method{format}{DATRASraw}(x)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
Dataframe in format x.
}
\description{
Format DATRAS data
}
