% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishglob.R
\name{formatFishglob}
\alias{formatFishglob}
\title{Format DATRAS data to FishGlob data set}
\usage{
formatFishglob(x)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
Dataframe in format of FishGlob data set.
}
\description{
Format DATRAS data to FishGlob data set
}
