% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getLongFormat}
\alias{getLongFormat}
\title{Create long-format table}
\usage{
getLongFormat(x)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
Long-format table
}
\description{
Create long-format table
}
