% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_prune.R
\name{list_prune_datras_available}
\alias{list_prune_datras_available}
\title{Available columns in a datras_raw object}
\usage{
list_prune_datras_available(x)
}
\arguments{
\item{x}{A \code{datras_raw} object.}
}
\value{
A named list with column names in \code{HH}, \code{HL}, and \code{CA}.
}
\description{
Available columns in a datras_raw object
}
