% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_prune.R
\name{list_prune_datras_defaults}
\alias{list_prune_datras_defaults}
\title{Default columns used by prune_datras}
\usage{
list_prune_datras_defaults()
}
\value{
A named list with default columns for \code{HH}, \code{HL}, and \code{CA}.
}
\description{
Default columns used by prune_datras
}
