% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini}
\alias{mini}
\title{Mini DATRAS survey data (example dataset)}
\format{
A list of class 'datras_raw' with 3 elements:
\describe{
\item{CA}{Biological data}
\item{HH}{Survey level information}
\item{HL}{Length measurements}
}
}
\source{
ICES DATRAS database \url{https://datras.ices.dk}
}
\usage{
mini
}
\description{
Example ICES DATRAS data for 6 years (2015-2020), 4 surveys (NS-IBTS, BITS,
BTS, and EVHOE), all quarters, and 5 species: Lophius piscatorius (European anglerfish),
Lepidorhombus whiffiagonis (Megrim), Hippoglossoides platessoides (American
plaice), Trisopterus esmarkii (Norway pout), and Amblyraja radiata (Starry
ray).
}
\details{
The years overlap with the public FishGlob data, so the data set is also
used to compare \code{\link[=clean_fishglob]{clean_fishglob()}} with FishGlob.
}
\keyword{datasets}
