% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHauls}
\alias{plotHauls}
\title{Plot hauls}
\usage{
plotHauls(plot.map = TRUE, xlim = NULL, ylim = NULL)
}
\arguments{
\item{plot.map}{TRUE}

\item{xlim}{NULL}

\item{ylim}{NULL}
}
\description{
Plot hauls
}
