% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSurveys}
\alias{plotSurveys}
\title{Plot surveys}
\usage{
plotSurveys(
  plot.map = TRUE,
  fixed.axes = TRUE,
  overlay = FALSE,
  consider.quarter = FALSE,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{plot.map}{TRUE}

\item{fixed.axes}{TRUE}

\item{overlay}{FALSE}

\item{consider.quarter}{FALSE}

\item{xlim}{NULL}

\item{ylim}{NULL}
}
\description{
Plot surveys
}
