% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prune}
\alias{prune}
\title{Prune data}
\usage{
prune(x, do.fishglob = FALSE)
}
\arguments{
\item{x}{a DATRASraw object.}

\item{do.fishglob}{FALSE}
}
\value{
Pruned DATRASraw object.
}
\description{
Prune data
}
