% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{readDATRAS}
\alias{readDATRAS}
\title{Download DATRAS survey information}
\usage{
readDATRAS(paths, years = NULL)
}
\arguments{
\item{surveys}{surveys}
}
\description{
Download DATRAS survey information
}
