% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{renameSurvey}
\alias{renameSurvey}
\title{Rename surveys in DATRASraw}
\usage{
renameSurvey(x, survNameOld, survNameNew)
}
\arguments{
\item{x}{a DATRASraw object.}

\item{survNameOld}{Old survey name}

\item{survNameNew}{New survey name}
}
\value{
DATRASraw with updated survey names
}
\description{
Rename surveys in DATRASraw
}
