% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawning_info}
\alias{spawning_info}
\title{Spawning information lookup table}
\format{
A data frame with the columns:
\describe{
\item{species}{Scientific name of the species.}
\item{aphia}{WoRMS AphiaID of the species.}
\item{ices_area}{ICES area the record applies to (e.g. 3.d.27).}
\item{spawn_months}{List column of integer months (1-12) in which the species spawns in that area.}
\item{source}{Source(s) the spawning information was compiled from (e.g. gofish, wkmat).}
\item{match_type}{How the area was matched to the source (e.g. exact_area, parent_area, neighbour_region).}
\item{source_area}{ICES area of the original source record used for the match.}
}
}
\source{
\url{https://github.com/federico-maioli/length_at_maturity}
}
\usage{
spawning_info
}
\description{
A lookup table giving the spawning months of fish species by ICES area,
compiled from the GoFish and WKMAT sources.
}
\keyword{datasets}
