% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_info_full_raw}
\alias{survey_info_full_raw}
\title{Full survey information table}
\format{
A data frame with columns:
\describe{
\item{Survey}{Survey code (e.g. NS-IBTS)}
\item{Year}{Years}
\item{Quarter}{Quarters}
\item{Gear}{Gear}
\item{StatRec}{ICES Statistical rectangle}
\item{lon}{Longitude}
\item{lat}{Latitude}
}
}
\source{
ICES DATRAS metadata.
}
\usage{
survey_info_full_raw
}
\description{
Extended version of \code{survey_info} with additional coverage and technical
details for each DATRAS survey.
}
\keyword{datasets}
