% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surveyInfo}
\alias{surveyInfo}
\title{Survey information table}
\format{
A data frame with columns:
\describe{
\item{survey}{Survey code (e.g. NS-IBTS)}
\item{years}{Years}
\item{quarters}{Quarters}
\item{description}{Description}
}
}
\source{
ICES DATRAS metadata.
}
\usage{
surveyInfo
}
\description{
Metadata for ICES bottom trawl surveys, including survey names, areas,
and available time ranges.
}
\keyword{datasets}
