% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolffish}
\alias{wolffish}
\title{Wolffish survey data (example dataset)}
\format{
A list of class 'datras_raw' with 3 elements:
\describe{
\item{CA}{Biological data}
\item{HH}{Survey level information}
\item{HL}{Length measurements}
}
}
\source{
ICES DATRAS database \url{https://datras.ices.dk}
}
\usage{
wolffish
}
\description{
Example ICES DATRAS data for \emph{Anarhichas lupus} (wolffish) from survey
NS-IBTS Q1 in years 1991-2024.
}
\keyword{datasets}
