% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{add_weight_at_length}
\alias{add_weight_at_length}
\title{Add weight-at-length estimates to a \code{datras_raw} object}
\usage{
add_weight_at_length(
  x,
  max_length = NULL,
  max_weight = NULL,
  plus_group = FALSE,
  lw_source = c("lookup", "ca"),
  lookup_as_backup = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{datras_raw} object.}

\item{max_length}{Optional numeric value giving the maximum length in
centimetres to retain when fitting the length-weight relationship.
Observations above this value are excluded.}

\item{max_weight}{Optional numeric value giving the maximum individual
weight in grams to retain when fitting the length-weight relationship.
Observations above this value are excluded.}

\item{plus_group}{Logical. If \code{TRUE} the midlength for the weight
calculation of the last length bin is not using the upper limit of this
length bin, which might be \code{Inf} or arbitrarily high and result in an
unrealistically high weight for that length bin. Instead the lower limit
plus half of the size of the second last length bin is used to define the
mid length of the largest length bin.}

\item{lw_source}{Character string specifying the source of length-weight
parameters. One of \code{"ca"} or \code{"lookup"}.}

\item{lookup_as_backup}{Logical. If \code{TRUE}, lookup parameters of the
length-weight relationship (a, b) in the species_info table are used.
Default: \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} (default), progress messages are printed.}
}
\value{
A \code{datras_raw} object with estimated weight information added.
}
\description{
Estimate catch weight from length data and add the resulting weight fields to
a \code{datras_raw} / \code{DATRASraw} object.
}
\details{
The function derives weight-at-length either from an lookup length-weight
relationship fitted to the \code{CA} table or, if \code{lw_source = "lookup"}, uses
length-weight parameters from the species_info table.

\itemize{
\item If \code{lw_source = "lookup"}, a linear model of the form
\eqn{\log(IndWgt) ~ \log(LngtCm)} is fitted to positive individual weights
in the \code{CA} table. Predicted weights are then assigned to length classes
defined by \code{attr(x, "cm.breaks")}, multiplied by numbers-at-length, and
optionally divided by haul duration.

\item If \code{lw_source = "lookup"}, weight is added using length-weight
parameters from the species_info table. }
}
\examples{

## Add numbers at length
dab <- add_numbers_at_length(dab)

## Add fitted weight-at-length estimates
x <- add_weight_at_length(dab)

## Exclude large values when fitting the length-weight model
x <- add_weight_at_length(dab, max_length = 100, max_weight = 10000)

## Use lookup weight-at-length instead from the species_info table
x <- add_weight_at_length(dab, lw_source = "lookup")

}
\seealso{
\code{\link[=check_weights]{check_weights()}}, \code{\link[=add_total_weight_by_haul]{add_total_weight_by_haul()}}
}
