% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{add_weight_at_length}
\alias{add_weight_at_length}
\title{Add weight-at-length estimates to a \code{datras_raw} object}
\usage{
add_weight_at_length(
  x,
  per_minute = TRUE,
  max_length = NULL,
  max_weight = NULL,
  plus_group = FALSE,
  empirical = FALSE
)
}
\arguments{
\item{x}{A \code{datras_raw} object.}

\item{per_minute}{Logical. If \code{TRUE} (default), estimated weights are
divided by haul duration in minutes.}

\item{max_length}{Optional numeric value giving the maximum length in
centimetres to retain when fitting the length-weight relationship.
Observations above this value are excluded.}

\item{max_weight}{Optional numeric value giving the maximum individual
weight in grams to retain when fitting the length-weight relationship.
Observations above this value are excluded.}

\item{plus_group}{Logical. If \code{TRUE} the midlength for the weight
calculation of the last length bin is not using the upper limit of this
length bin, which might be \code{Inf} or arbitrarily high and result in an
unrealistically high weight for that length bin. Instead the lower limit
plus half of the size of the second last length bin is used to define the
mid length of the largest length bin.}

\item{empirical}{Logical. If \code{TRUE}, use empirical weight-at-length
calculations instead of fitting a length-weight model to the \code{CA} table.}
}
\value{
A \code{datras_raw} object with estimated weight information added.
}
\description{
Estimate catch weight from length data and add the resulting weight fields to
a \code{datras_raw} / \code{DATRASraw} object.
}
\details{
The function derives weight-at-length either from an empirical
length-weight relationship fitted to the \code{CA} table or, if
\code{empirical = TRUE}, uses length-weight parameters from the species_info table.

\itemize{
\item If \code{empirical = FALSE}, a linear model of the form
\eqn{\log(IndWgt) ~ \log(LngtCm)} is fitted to positive individual weights
in the \code{CA} table. Predicted weights are then assigned to length classes
defined by \code{attr(x, "cm.breaks")}, multiplied by numbers-at-length, and
optionally divided by haul duration.

\item If \code{empirical = TRUE}, weight is added using length-weight
parameters from the species_info table. }
}
\examples{

## Add numbers at length
dab <- add_numbers_at_length(dab)

## Add fitted weight-at-length estimates
x <- add_weight_at_length(dab)

## Exclude large values when fitting the length-weight model
x <- add_weight_at_length(dab, max_length = 100, max_weight = 10000)

## Use empirical weight-at-length instead
x <- add_weight_at_length(dab, empirical = TRUE)

}
\seealso{
\code{\link[=check_weights]{check_weights()}}, \code{\link[=add_total_weight_by_haul]{add_total_weight_by_haul()}}
}
