% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swept_area.R
\name{addSweptAreaFishGlob}
\alias{addSweptAreaFishGlob}
\title{Add Swept Area index following the FishGlob calculations}
\usage{
addSweptAreaFishGlob(x)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
DATRASraw object with SweptArea index.
}
\description{
Add Swept Area index following the FishGlob calculations
}
