% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishglob.R
\name{addWeightFishglob}
\alias{addWeightFishglob}
\title{Calculate weight by length classes using empirical a and b, and add
to HH-records}
\usage{
addWeightFishglob(d, to1min = TRUE)
}
\arguments{
\item{d}{DATRASraw object}

\item{to1min}{divide by haul duration in minutes? (defaults to TRUE)}
}
\value{
DATRASraw object
}
\description{
Calculate weight by length classes using empirical a and b, and add to
HH-records accoring to FishGlob workflow
}
