% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{as_long_format}
\alias{as_long_format}
\title{Create long-format table}
\usage{
as_long_format(x)
}
\arguments{
\item{x}{a datras_raw object.}
}
\value{
Long-format table
}
\description{
Create long-format table
}
