% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean data}
\usage{
clean(
  x,
  aphias = NULL,
  years = NULL,
  quarters = NULL,
  gears = NULL,
  impute.missing.depth = TRUE,
  correct.species = TRUE,
  do.fishglob = FALSE
)
}
\arguments{
\item{x}{a DATRASraw object.}

\item{aphias}{x}

\item{years}{x}

\item{quarters}{x}

\item{gears}{x}

\item{impute.missing.depth}{TRUE}

\item{correct.species}{TRUE}

\item{do.fishglob}{FALSE}
}
\value{
Cleaned DATRASraw object.
}
\description{
Clean data
}
