% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishglob.R
\name{cleanFishglob}
\alias{cleanFishglob}
\title{Clean data following FishGlob workflow}
\usage{
cleanFishglob(x)
}
\arguments{
\item{x}{a DATRASraw object.}

\item{aphias}{x}

\item{years}{x}

\item{quarters}{x}

\item{gears}{x}
}
\value{
Cleaned DATRASraw object according to FishGlob workflow.
}
\description{
This function follows the workflow that was used to create the
FishGlob data set (Maureaud et al. 2021). Adjusted from:
\url{https://github.com/fishglob/FishGlob_data/tree/main/cleaning_codes}.

A Maureaud, A., Frelat, R., Pécuchet, L., Shackell, N., Mérigot, B., Pinsky,
M.L., Amador, K., Anderson, S.C., Arkhipkin, A., Auber, A. and Barri, I.,
2021. Are we ready to track climate-driven shifts in marine species across
international boundaries? A global survey of scientific bottom trawl data.
Global change biology, 27(2), pp.220-236.
}
