% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dab}
\alias{dab}
\title{Dab survey data (example dataset)}
\format{
A list of class 'DATRASraw' with 3 elements:
\describe{
\item{CA}{Biological data}
\item{HH}{Survey level information}
\item{HL}{Length measurements}
}
}
\source{
ICES DATRAS database \url{https://datras.ices.dk}
}
\usage{
dab
}
\description{
Example ICES DATRAS data for \emph{Limanda limanda} (dab) from survey NS-IBTS in
years 2020-2023.
}
\keyword{datasets}
