% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{package}
\name{DATRASextra-package}
\alias{DATRASextra}
\alias{DATRASextra-package}
\title{DATRASextra package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to download, read, clean, check, summarise and visualise bottom-trawl survey data from the ICES Database of Trawl Surveys (DATRAS, \url{https://datras.ices.dk}). Downloads are archived as exchange files with a manifest, so that an extraction can be verified and reproduced, and are read into one object with derived variables such as haul identifiers, lengths in centimetres and raised numbers at length. Further functions add swept area, numbers and weight at length, ICES areas and depth, flag implausible values, compute stratified abundance indices and spatial indicators, build prediction grids for spatial models, and export tidy tables, including the format of the FishGlob database. The reader and the data class are adapted from the 'DATRAS' package by Kasper Kristensen and Casper W. Berg.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tokami.github.io/DATRASextra/}
  \item \url{https://github.com/tokami/DATRASextra}
  \item Report bugs at \url{https://github.com/tokami/DATRASextra/issues}
}

}
\author{
\strong{Maintainer}: Tobias Mildenberger \email{tobm@dtu.dk} (\href{https://orcid.org/0000-0002-6631-7524}{ORCID})

Authors:
\itemize{
  \item Tobias Mildenberger \email{tobm@dtu.dk} (\href{https://orcid.org/0000-0002-6631-7524}{ORCID})
  \item Federico Maioli \email{fedma@aqua.dtu.dk} (\href{https://orcid.org/0000-0002-8305-1609}{ORCID})
  \item Casper Berg \email{cbe@aqua.dtu.dk} (\href{https://orcid.org/0000-0002-3812-5269}{ORCID})
}

}
\keyword{internal}
