% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{package}
\name{DATRASextra-package}
\alias{DATRASextra}
\alias{DATRASextra-package}
\title{DATRASextra package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

DATRASextra is an R package for working with ICES DATRAS survey data, with tools for downloading, processing, checking, summarising, and visualising bottom-trawl survey data. The package extends the functionality of DATRAS with additional data workflows, quality checks, derived variables and indicators, plotting tools, and analysis-ready outputs. It is intended to support reproducible research, stock assessment, and survey-based analyses in fisheries science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tokami.github.io/DATRASextra/}
  \item \url{https://github.com/tokami/DATRASextra}
  \item Report bugs at \url{https://github.com/tokami/DATRASextra/issues}
}

}
\author{
\strong{Maintainer}: Tobias Mildenberger \email{tobm@dtu.dk} (\href{https://orcid.org/0000-0002-6631-7524}{ORCID})

Authors:
\itemize{
  \item Tobias Mildenberger \email{tobm@dtu.dk} (\href{https://orcid.org/0000-0002-6631-7524}{ORCID})
  \item Federico Maioli \email{fedma@aqua.dtu.dk} (\href{https://orcid.org/0000-0002-8305-1609}{ORCID})
  \item Casper Berg \email{cbe@aqua.dtu.dk} (\href{https://orcid.org/0000-0002-3812-5269}{ORCID})
}

}
\keyword{internal}
