% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_land}
\alias{get_land}
\title{Load dissolved Natural Earth land polygon}
\source{
Natural Earth via rnaturalearth.
}
\usage{
get_land(download_map = FALSE, scale = 50)
}
\arguments{
\item{download_map}{Logical. If \code{FALSE} the download map included in
DATRASextra is used for plotting (with 1:50m scale). If \code{TRUE}, a map is
downloaded with \code{\link[rnaturalearth:ne_download]{rnaturalearth::ne_download()}} and the specified scale.}

\item{scale}{Scale of the map to be downloaded by
\code{\link[rnaturalearth:ne_download]{rnaturalearth::ne_download()}}. (Only used if download_map = TRUE).}
}
\value{
An sf object (single MULTIPOLYGON) in EPSG:4326.
}
\description{
If download_map = FALSE, loads a dissolved land polygon shipped with the
package (Natural Earth 1:50m), used for offline plotting.
}
