% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini}
\alias{mini}
\title{Mini DATRAS survey data (example dataset)}
\format{
A list of class 'DATRASraw' with 3 elements:
\describe{
\item{CA}{Biological data}
\item{HH}{Survey level information}
\item{HL}{Length measurements}
}
}
\source{
ICES DATRAS database \url{https://datras.ices.dk}
}
\usage{
mini
}
\description{
Example ICES DATRAS data for 2 years (2022-2023), 4 surveys (NS-IBTS, BITS,
BTS, and EVHOE), and 5 species: Lophius piscatorius (European anglerfish),
Lepidorhombus whiffiagonis (Megrim), Hippoglossoides platessoides (American
plaice), Trisopterus esmarkii (Norway pout), and Amblyraja radiata (Starry
ray).
}
\keyword{datasets}
