% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_fishglob}
\alias{mini_fishglob}
\title{Mini DATRAS survey data for fishglob comparison (example dataset)}
\format{
A list of class 'datras_raw' with 3 elements:
\describe{
\item{CA}{Biological data}
\item{HH}{Survey level information}
\item{HL}{Length measurements}
}
}
\source{
ICES DATRAS database \url{https://datras.ices.dk}
}
\usage{
mini_fishglob
}
\description{
Example ICES DATRAS data for 5 years (2015-2020), 4 surveys (NS-IBTS, BITS,
BTS, and EVHOE), and 5 species: Lophius piscatorius (European anglerfish),
Lepidorhombus whiffiagonis (Megrim), Hippoglossoides platessoides (American
plaice), Trisopterus esmarkii (Norway pout), and Amblyraja radiata (Starry
ray).
}
\keyword{datasets}
