% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHaulsBySurvey}
\alias{plotHaulsBySurvey}
\title{Plot hauls by survey}
\usage{
plotHaulsBySurvey(
  plot.map = TRUE,
  fixed.scale = TRUE,
  col = hcl.colors(12, "YlOrRd", rev = TRUE),
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{plot.map}{TRUE}

\item{fixed.scale}{TRUE}

\item{col}{colours}

\item{xlim}{NULL}

\item{ylim}{NULL}
}
\description{
Plot hauls by survey
}
