% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSurveys}
\alias{plotSurveys}
\title{Plot surveys}
\usage{
plotSurveys(
  plot.map = TRUE,
  fixed.axes = TRUE,
  overlay = FALSE,
  consider.quarter = FALSE,
  xlim = NULL,
  ylim = NULL
)
}
\description{
Plot surveys
}
