% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishglob.R
\name{prune_fishglob}
\alias{prune_fishglob}
\title{Prune a \code{datras_raw} object following the FishGlob workflow}
\usage{
prune_fishglob(x)
}
\arguments{
\item{x}{A \code{datras_raw} object.}
}
\value{
A pruned \code{datras_raw} object following the FishGlob workflow.
}
\description{
Reduce a \code{datras_raw} / \code{DATRASraw} object to the components and columns
needed for the FishGlob workflow.
}
\details{
This function removes the \code{CA} table entirely and retains only a selected
subset of columns in the \code{HH} and \code{HL} tables. It is mainly intended to
reduce memory use when preparing data in a format consistent with the
FishGlob data workflow.

The FishGlob workflow does not require the full DATRAS object. In particular,
the \code{CA} table is discarded and only a reduced set of variables is kept from
the \code{HH} and \code{HL} tables.

This can substantially reduce memory use when working with large survey data
sets.
}
\examples{
\dontrun{
## as prune_datras(), but with the fixed column set of the FishGlob
## workflow; the CA table is dropped entirely
x_small <- prune_fishglob(x)
}

}
\seealso{
\code{\link[=prune_datras]{prune_datras()}}, \code{\link[=clean_fishglob]{clean_fishglob()}}
}
