% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishglob.R
\name{pruneFishglob}
\alias{pruneFishglob}
\title{Prune data according to FishGlob workflow}
\usage{
pruneFishglob(x)
}
\arguments{
\item{x}{a DATRASraw object.}
}
\value{
Pruned DATRASraw object according to FishGlb workflow.
}
\description{
Prune data according to FishGlob workflow
}
\details{
DATRAS' CA data set, and some columns in the 'HH' and 'HL' data
sets are not needed to reproduce the FishGlob data set. Thus, these data
sets and columns can be removed to save some memory.
}
