% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{writeExchange}
\alias{writeExchange}
\title{Write DATRASraw to Exchange}
\usage{
writeExchange(x, zipfile = "DATRAS.zip")
}
\arguments{
\item{x}{a DATRASraw object}
}
\description{
Write DATRASraw to Exchange
}
